Bugfixing Click and Drag #1
Click and Drag has had an issue fixed where the object would be able to be dragged out of the goal yet be unable to move. Also fixed an issue with the Z co-ordinate being changed meaning object order could not be set. Currently, dragged object stops moving as soon as it hits the goal. This will be fixed in a future update.